21_job_tracker_fullstack/backend/requirements.txt
fastapi
uvicorn[pure]
sqlmodel
sqlalchemy>=2.0
psycopg2-binary
pydantic
python-dotenv
passlib[bcrypt]
pyjwt
python-multipart
